Criar Novo Prato Loja Controller

+++ Pitangueiros.GuardioesDasQuentinhas.Domain/Porcao.cs
@@ -7,7 +7,7 @@ namespace Pitangueiros.GuardioesDasQuentinhas.Domain.Entities {
     public class Porcao : EntidadeBase<long>, IDeleteLogico, IAuditoria{
         public string Item { get; set; }
         public int Quantidade { get; set; }
-        public int Preco { get; set; }
+        public double Preco { get; set; }
         public CategoriaPorcao CategoriaPorcao { get; set; }
         public bool IsDeleted { get; set; }
         public DateTime DataCriacao { get; set; }
